+++ source/Detector.cpp
@@ -152,7 +152,8 @@ Detector::Detector(ConfigurationParameters &configParam, HDF5File &hdf5file, Cam
   includeCTIeffects(true), 
   includeOpenShutterSmearing(true), 
   includeQuantumEfficiency(true),
-  includeVignetting(true),
+  includeNaturalVignetting(true),
+  includeMechanicalVignetting(true),
   includeParticulateContamination(true),
   includeMolecularContamination(true),
   includeFullWellSaturation(true),
@@ -185,6 +186,28 @@ Detector::Detector(ConfigurationParameters &configParam, HDF5File &hdf5file, Cam
 
     smearingMap.zeros(numRowsSmearingMap, numColumnsPixelMap);
     throughputMap.ones(numRowsPixelMap, numColumnsPixelMap);
+
+    // If we are going to apply open-shutter smearing, we have to know which pixels are within
+    // the FOV (relevant only in case of mechanical vignetting).  When mechanical vignetting is
+    // disabled, all pixels of the detector are inside the FOV.
+
+    if(includeOpenShutterSmearing)
+    {
+        // Mechanical vignetting map:
+        //  - no mechanical vignetting: all pixels of the sub-field inside FOV -> all values set to one
+        //  - mechanical vignetting: set value of the pixels in the sub-field outside FOV to zero (others should be one) -> on creation of the throughput map
+
+        mechanicalVignettingMask.ones(numRowsPixelMap, numColumnsPixelMap);
+
+        // Number of exposed rows in each column:
+        // - no mechanical vignetting: all exposed rows inside FOV (numRows - firstRowExposed)
+        // - mechanical vignetting: count the exposed rows (i.e. from firstRowExposed) that are inside FOV
+
+        numExposedRowsInFOV.zeros(numColumnsPixelMap);
+
+        if(!includeMechanicalVignetting)
+            numExposedRowsInFOV.fill(numRows - firstRowExposed);
+    }
 
     // Check whether the gain values for left- and right-hand side of the CCD are not too far apart
 
@@ -239,8 +262,6 @@ Detector::~Detector()
  *        value at the given time point
  *
  * \param time: current time
- *
- * \return 
  */
 
 void Detector::updateParameters(double time)
@@ -262,7 +283,7 @@ void Detector::updateParameters(double time)
  * \brief Configure the Detector object using the ConfigurationParameters
  * 
  * \param configParam: Configuration parameters
- **/
+ */
 
  void Detector::configure(ConfigurationParameters &configParam)
  {
@@ -331,7 +352,8 @@ void Detector::updateParameters(double time)
     fullWellSaturationLimit             = configParam.getLong("CCD/FullWellSaturation");
     digitalSaturationLimit              = configParam.getLong("CCD/DigitalSaturation");
     readoutNoise                        = configParam.getDouble("CCD/ReadoutNoise");
-    expectedValueVignetting             = configParam.getDouble("CCD/Vignetting/ExpectedValue");
+    expectedValueNaturalVignetting      = configParam.getDouble("CCD/Vignetting/NaturalVignetting/ExpectedValue");
+    radiusFOV                           = deg2rad(configParam.getDouble("CCD/Vignetting/MechanicalVignetting/RadiusFOV"));
     particulateContaminationEfficiency  = configParam.getDouble("CCD/Contamination/ParticulateContaminationEfficiency");
     molecularContaminationEfficiency    = configParam.getDouble("CCD/Contamination/MolecularContaminationEfficiency");
 
@@ -391,7 +413,8 @@ void Detector::updateParameters(double time)
     includeCTIeffects               = configParam.getBoolean("CCD/IncludeCTIeffects");
     includeOpenShutterSmearing      = configParam.getBoolean("CCD/IncludeOpenShutterSmearing");
     includeQuantumEfficiency        = configParam.getBoolean("CCD/IncludeQuantumEfficiency");
-    includeVignetting               = configParam.getBoolean("CCD/IncludeVignetting");
+    includeNaturalVignetting        = configParam.getBoolean("CCD/IncludeNaturalVignetting");
+    includeMechanicalVignetting     = configParam.getBoolean("CCD/IncludeMechanicalVignetting");
     includePolarization             = configParam.getBoolean("CCD/IncludePolarization");
     includeFullWellSaturation       = configParam.getBoolean("CCD/IncludeFullWellSaturation");
     includeDigitalSaturation        = configParam.getBoolean("CCD/IncludeDigitalSaturation");
@@ -406,6 +429,10 @@ void Detector::updateParameters(double time)
     numRowsBiasMap          = configParam.getInteger("SubField/NumBiasPrescanRows");
     numColumnsBiasMap       = configParam.getInteger("SubField/NumBiasPrescanColumns");
     numRowsSmearingMap      = configParam.getInteger("SubField/NumSmearingOverscanRows");
+
+
+
+    // No parallel over-scan in case of partial readout
 
     if(readoutMode == "Partial")
     {
@@ -428,118 +455,6 @@ void Detector::updateParameters(double time)
     numEdgePixels = 0;
  }
 
-
-
-
-
-
-
-// /**
-//  * \brief Configure the readout times for Detector object using the ConfigurationParameters
-//  *
-//  * \param configParam: Configuration parameters
-//  **/
-//void Detector::configureReadoutTime(ConfigurationParameters &configParam)
-//{
-//	// Both detector halves are read out simultaneously
-//	// -> columns read out by the FEE:
-//	// 		- half of the CCD
-//	// 		- serial pre-scan
-//	// 		- (serial over-scan)
-//
-//	int numColumnsReadout = numColumns / 2 + numColumnsBiasMap; // + numRowsSerialOverScan
-//
-//	// How many rows will be actually read out by the FEE?
-//	// 	- nominal mode: image area + parallel over-scan
-//	//      normal camera: image area = whole CCD
-//	//      fast camera: image area = lower half of the CCD
-//	//	- partial readout: configurable
-//	// The rest of the image area will be dumped
-//
-//	int numRowsReadout, numRowsDump;
-//
-//	// -----------
-//	// Fast camera
-//	// -----------
-//
-//	if (isFastCamera) {
-//
-//		// Move the upper half of the CCD down to the lower half, row-by-row
-//
-//		int numRowsFrameTransfer = numRows - firstRowExposed;
-//
-//		readoutTimeBeforeNextExposure = numRowsFrameTransfer
-//						* parallelTransferTimeFast;
-//
-//		// The actual readout of the lower half of the CCD (after frame transfer) is done
-//		// while the next exposure has already started
-//
-//		// Nominal mode
-//
-//		if (readoutMode == "Nominal")
-//		{
-//			numRowsReadout = firstRowExposed + numRowsSmearingMap;
-//			numRowsDump = 0;
-//
-//		}
-//
-//		// Partial readout
-//
-//		else if (readoutMode == "Partial")
-//		{
-//			numRowsReadout = numRowsPartialReadout;
-//			numRowsDump = firstRowExposed - numRowsReadout;
-//		}
-//
-//
-//		readoutTimeDuringNextExposure = numRowsDump * parallelTransferTimeFast
-//				+ numRowsReadout * (parallelTransferTime + numColumnsReadout * serialTransferTime);
-//	}
-//
-//	// -------------
-//	// Normal camera
-//	// -------------
-//
-//	else
-//	{
-//
-//		// Nominal mode (full-frame readout)
-//
-//		if (readoutMode == "Nominal")
-//		{
-//
-//			// Rows read out by the FEE:
-//			// 		- rows of image area
-//			// 		- parallel over-scan
-//
-//			numRowsReadout = numRows + numRowsSmearingMap;
-//
-//			// No rows dumped
-//
-//			numRowsDump = 0;
-//		}
-//
-//		// Partial readout
-//
-//		else if (readoutMode == "Partial") {
-//
-//			// Rows read out by the FEE: rows in the block (other rows in image area are dumped)
-//			// Note: no parallel over-scan
-//
-//			numRowsReadout = numRowsPartialReadout;
-//			numRowsDump = numRows - numRowsReadout;
-//
-//		}
-//
-//		readoutTimeBeforeNextExposure =
-//				numRowsReadout
-//						* (numColumnsReadout * serialTransferTime
-//								+ parallelTransferTime)
-//						+ numRowsDump * parallelTransferTimeFast;
-//
-//		readoutTimeDuringNextExposure = 0;
-//	}
-// }
 
 
 
@@ -630,6 +545,9 @@ void Detector::generateThroughputMap()
 
     throughputMap.fill(1.0);
 
+    if(includeMechanicalVignetting  && includeOpenShutterSmearing)
+        mechanicalVignettingMask.fill(1);
+
     double xFPmm, yFPmm;
     double angle;
 
@@ -639,7 +557,7 @@ void Detector::generateThroughputMap()
 //    const double refAngleQuantumEfficiencyRadians = deg2rad(refAngleQE);     // Reference angle for the quantum efficiency [radians]
 //    const double acosQuantumEfficiency = acos(relativeRefEfficiencyQE);        // Relative efficiency due to the angle dependency of the QE at the reference angle
 
-    if (includeVignetting || includePolarization || includeQuantumEfficiency)
+    if (includeNaturalVignetting || includeMechanicalVignetting || includePolarization || includeQuantumEfficiency)
     {
         // Loop over all pixels in the pixel map
 
@@ -655,9 +573,23 @@ void Detector::generateThroughputMap()
 
                 angle = camera.getGnomonicRadialDistanceFromOpticalAxis(xFPmm, yFPmm);
 
-                // Vignetting
+                // Mechanical vignetting
 
-                if (includeVignetting) throughputMap(row, column) *= pow(cos(angle), 2);
+                if(includeMechanicalVignetting)
+                {
+                    if (angle > radiusFOV)
+                    {
+                        throughputMap(row, column) = 0.0;
+
+                        if(includeOpenShutterSmearing)
+                            mechanicalVignettingMask(row, column) = 0;
+                    }
+                }
+
+                // Natural vignetting
+
+                if (includeNaturalVignetting) 
+                    throughputMap(row, column) *= pow(cos(angle), 2);
 
                 // Polarisation (Eq. 4-11 in PLATO-DLR-PL-RP-001)
 
@@ -1278,9 +1210,9 @@ void Detector::addPhotonNoise()
 void Detector::addCosmics(float exposureTime)
 {
 	cosmicHitRateDistribution     = poisson_distribution<long>(cosmicHitRate);                                       // [hits/cm^2/s]
-    cosmicEntryColumnDistribution = uniform_real_distribution<double>(0, numColumnsPixelMap - 1);                    // []
-    cosmicEntryAngleDistribution  = uniform_real_distribution<double>(0, 2 * PI);                                    // [rad]
-    cosmicTrailLengthDistribution = uniform_real_distribution<double>(cosmicTrailLength[0], cosmicTrailLength[1]);   // [pixel]
+    cosmicEntryColumnDistribution = uniform_real_distribution<double>(0, numColumnsPixelMap - 1);                    // [pixels]
+    cosmicEntryAngleDistribution  = uniform_real_distribution<double>(0, 2 * PI);                                    // [radians]
+    cosmicTrailLengthDistribution = uniform_real_distribution<double>(cosmicTrailLength[0], cosmicTrailLength[1]);   // [pixels]
     cosmicIntensityDistribution   = uniform_real_distribution<double>(cosmicIntensity[0], cosmicIntensity[1]);       // [e-/hit]
 
     // Cosmics in the subfield
@@ -1866,20 +1798,75 @@ void Detector::applyShort2013CTImodel()
  */
 void Detector::applyOpenShutterSmearing(float exposureTime)
 {
+    if (includeMechanicalVignetting)
+    {
+        // The mask indicating which pixels in the sub-field are within the FOV and which
+        // ones not, has already been created upon construction of the throughput map.
+
+        double rowFOV, xFPmmFirstExposed, yFPmmFirstExposed, xFPmmTop, yFPmmTop, angleFirstExposed, angleTop;
+
+        // Loop over all columns in the sub-field
+
+        for(unsigned int column = 0; column < numColumnsPixelMap; column++) //[pixels in sub-field]
+        {
+            // Intersection of the current column of the detector with the circle representing the FOV:
+            // (rowFOV, column).  If no intersection can be found, this is NaN.
+            rowFOV = getRowEdgeFOV(column);
+
+            if(isnan(rowFOV))
+            {
+                // All exposed rows are within the FOV
+
+                numExposedRowsInFOV(column) = numRows - firstRowExposed;
+            }
+            else
+            {
+                tie(xFPmmFirstExposed, yFPmmFirstExposed) = pixelToFocalPlaneCoordinates(firstRowExposed, column + subFieldZeroPointColumn);    // detector [pixels] -> focal plane [mm]
+                angleFirstExposed = camera.getGnomonicRadialDistanceFromOpticalAxis(xFPmmFirstExposed, yFPmmFirstExposed);
+
+                tie(xFPmmTop, yFPmmTop) = pixelToFocalPlaneCoordinates(numRows - 1, column + subFieldZeroPointColumn);
+                angleTop = camera.getGnomonicRadialDistanceFromOpticalAxis(xFPmmTop, yFPmmTop);
+
+                if (angleFirstExposed > radiusFOV)
+                {
+                    if (angleTop > radiusFOV)
+                        numExposedRowsInFOV(column) = 0;
+                    
+                    else
+                        numExposedRowsInFOV(column) = numRows - (int) rowFOV;
+                }
+
+                else
+                {
+                    if (angleTop > radiusFOV)
+                        numExposedRowsInFOV(column) = (int) rowFOV - firstRowExposed + 1;
+                    else
+                        numExposedRowsInFOV(column) = numRows - firstRowExposed;
+                }
+            }
+        }
+    }
+
+
+
     // Average out the fluxes in the pixel map per column (of the whole CCD) and make sure it is
-    // scaled with the readout time instead of with the exposure time:
-    // - rows in the sub-field: use actual fluxes
-    // - rows outside the sub-field: use total sky background
+    // scaled with the readout time (during which the detector is susceptible to open-shutter smearing) 
+    // instead of with the exposure time:
+    // - rows in the sub-field (in the exposed part of the detector and in the FOV): use actual fluxes (accumulated during exposure)
+    // - rows outside the sub-field (in the exposed part of the detector and in the FOV): use total sky background
+    // - rows outside the FOV and/or not exposed are not considered (as these rows are shielded off against incoming radiation)
 
-    arma::Row<float> openShutterSmearing = arma::sum(pixelMap, 0);
+    arma::Row<float> openShutterSmearing = arma::sum(pixelMap % mechanicalVignettingMask, 0);   // Flux in the exposed part of the sub-field that is inside the FOV (per column)
+    arma::Row<int> numExposedSubFieldRowsInFOV = arma::sum(mechanicalVignettingMask, 0);        // Number of pixels in the exposed part of the sub-field that are inside the FOV (per column)
 
-    int numRowsExposedDuringReadout = numRows - firstRowExposed - numRowsPixelMap + numRowsSmearingMap;
-    float openShutterSmearingOutsideSubField = camera.getTotalSkyBackground() * numRowsExposedDuringReadout;
+    arma::Row<int> numExposedNonSubFieldRowShifts = numExposedRowsInFOV - numExposedSubFieldRowsInFOV + numRowsSmearingMap; // Number of pixels in the exposed part of the detector that do not reside in the sub-field + parallel over-scan
+    arma::Row<float> openShutterSmearingOutsideSubField = arma::conv_to<arma::Row<float>>::from(numExposedNonSubFieldRowShifts) * camera.getTotalSkyBackground();
 
-    // Apply all throughput efficiencies
+   // Apply all throughput efficiencies (these have not been applied to the total sky background yet)
+   // Note that mechanical vignetting has already been taken into account (if applicable)
 
-    if(includeVignetting)
-        openShutterSmearingOutsideSubField *= expectedValueVignetting;
+    if(includeNaturalVignetting)
+        openShutterSmearingOutsideSubField *= expectedValueNaturalVignetting;
 
     if(includePolarization)
         openShutterSmearingOutsideSubField *= expectedValuePolarization;
@@ -1896,14 +1883,16 @@ void Detector::applyOpenShutterSmearing(float exposureTime)
     if(includeQuantumEfficiency)
         openShutterSmearingOutsideSubField *= meanQE;
 
-
     openShutterSmearing += openShutterSmearingOutsideSubField;
 
     // Fast camera: lower half of the CCD is shielded off -> no contribution to open-shutter smearing here
 
-    float factor = readoutTimeBeforeNextExposure / exposureTime / (numRows - firstRowExposed);
+    arma::Row<float> openShutterSmearingTime = arma::conv_to<arma::Row<float>>::from(numExposedRowsInFOV + numRowsSmearingMap) / 
+        (numRows - firstRowExposed + numRowsSmearingMap) * readoutTimeBeforeNextExposure;
 
-    openShutterSmearing *= factor;
+    arma::Row<float> factor = openShutterSmearingTime % arma::pow(arma::conv_to<arma::Row<float>>::from(numExposedRowsInFOV), -1) / exposureTime;
+
+    openShutterSmearing %= factor;
 
     // Add the effect of the open-shutter smearing to the pixel map
 
@@ -1914,6 +1903,68 @@ void Detector::applyOpenShutterSmearing(float exposureTime)
 
     for (unsigned int row = 0; row < numRowsSmearingMap; row++)
         smearingMap(row, arma::span::all) += openShutterSmearing;
+}
+
+
+
+
+
+
+
+
+
+
+/**
+ * \brief Find the intersection of the circle representing the FOV and the given column of the
+ *        detector.  In case this circle does intersect with the detector, the row coordinate
+ *        is returned; if not, NaN is returned.
+ * 
+ * \param column: Column (coordinate) [pixels], for which to find the intersection with the
+ *                circle representing the FOV.
+ * 
+ * \return In case this circle does intersect with the detector, the row coordinate
+ *         is returned; if not, NaN is returned.
+ **/ 
+double Detector::getRowEdgeFOV(int column)
+{
+    double pixelSizeMm = pixelSize / 1000.0;    // Pixel size [µm] -> [mm]
+
+    // Quadratic equation: a * x**2 + b * x + c  = 0
+    // Find intersection between circle representing the FOV and the given column
+
+    double a = pow(pixelSizeMm, 2);
+    double b = - 2 * pixelSizeMm * subFieldZeroPointRow;
+    double c = pow(subFieldZeroPointRow, 2) + pow(column * pixelSizeMm - subFieldZeroPointColumn , 2) -  pow(camera.getFocalLength() * tan(radiusFOV), 2);
+
+    // Discriminant (should be positive)
+
+    double discriminant = pow(b, 2) - 4 * a * c;
+
+    if (discriminant < 0)
+        return nan("");
+    
+    double discriminantSqrt = sqrt(discriminant);
+
+    // First solution of the equation
+    // Check whether it intersects with the detector (if not, try the other solution)
+
+    double solutionRow = (-b + discriminantSqrt) / (2 * a);
+
+    if ((solutionRow >= 0) && (solutionRow < numRows))
+        return solutionRow;
+
+    // Second solution of the equation
+    // Check whther it intersects with the detector (if not, the whole row is within the FOV)
+
+    solutionRow = (-b - discriminantSqrt) / (2 * a);
+    
+    if ((solutionRow >= 0) && (solutionRow < numRows))
+        return solutionRow;
+
+    // No proper solution found (i.e. circle of FOV does not intersect
+    // with the detector in the given column)
+
+    return nan("");
 }
 
 
@@ -2047,7 +2098,7 @@ void Detector::applyQuantisation()
     if (includeDigitalSaturation)
     {
         Log.debug("Detector: applying digital saturation to pixelMap, biasMap and smearingMap (digitalSaturationLimit=" + to_string(digitalSaturationLimit) + ")");
-            applyDigitalSaturation();
+        applyDigitalSaturation();
     }
     else
     {
@@ -2172,7 +2223,7 @@ void Detector::addElectronicOffset()
 /**
  * \brief: Apply the effect of digital saturation to the pixel map,
  *         smearing map, and bias register map. This means that the pixel values in
- *         these maps (expressed in [ADU / pixel]) are topped off to the digital saturation
+ *         these maps (expressed in [ADU / pixel]) are Lastped off to the digital saturation
  *         limit of the detector (also expressed in [ADU / pixel]).
  *
  * \pre Pixel unit in the pixel, smearing, and bias maps: [ADU].
@@ -2182,16 +2233,16 @@ void Detector::addElectronicOffset()
  */
 void Detector::applyDigitalSaturation()
 {
-    // Top off the values in the pixel map
+    // Last off the values in the pixel map
 
     pixelMap(arma::find(pixelMap > digitalSaturationLimit)).fill(digitalSaturationLimit);
 
-    // Top off the values in the bias register map
+    // Last off the values in the bias register map
 
     biasMapLeft(arma::find(biasMapLeft > digitalSaturationLimit)).fill(digitalSaturationLimit);
     biasMapRight(arma::find(biasMapRight > digitalSaturationLimit)).fill(digitalSaturationLimit);
 
-    // Top off the values in the smearing map
+    // Last off the values in the smearing map
 
     smearingMap(arma::find(smearingMap > digitalSaturationLimit)).fill(digitalSaturationLimit);
 }
@@ -2511,7 +2562,6 @@ void Detector::initHDF5Groups()
  * 
  * \param exposureNr:   Sequential number of the exposure
  */
-
 void Detector::writePixelMapsToHDF5(int exposureNr)
 {
 	// Compose the image name