+++ CHANGELOG.md
@@ -29,7 +29,13 @@
 
 * Incorrect flatfield level (GitHub issue #326)
 
-* Segfault on running test harness (GitHub issue # 329)
+* Segfault on running test harness (GitHub issue #329)
+
+* Open-shutter smearing not accounted for (GitHub issue #339)
+
+* Implement mechanical vignetting (GitHub issue #334)
+
+* Partial-readout parameters not read out correctly (GitHub issue #346)